Generate part of module in separate file and only loaded if needed
